--- a/arangod/Aql/AqlItemMatrix.cpp
+++ b/arangod/Aql/AqlItemMatrix.cpp
@@ -46,6 +46,16 @@ std::pair<SharedAqlItemBlockPtr, size_t> AqlItemMatrix::getBlock(size_t index) c
   return  {_blocks[index], index == 0 ? _startIndexInFirstBlock : 0};
 }
 
+std::pair<AqlItemBlock const*, size_t> AqlItemMatrix::getBlockRef(size_t index) const noexcept {
+  TRI_ASSERT(index < numberOfBlocks());
+  // The first block could contain a shadowRow
+  // and the first unused data row, could be after the
+  // shadowRow.
+  // All other blocks start with the first row as data row
+  TRI_ASSERT(_blocks[index].get());
+  return  {_blocks[index].get(), index == 0 ? _startIndexInFirstBlock : 0};
+}
+
 InputAqlItemRow AqlItemMatrix::getRow(AqlItemMatrix::RowIndex index) const noexcept {
   auto const& [block, unused] = getBlock(index.first);
   TRI_ASSERT(index.second >= unused);