--- a/arangod/Aql/AqlItemMatrix.h
+++ b/arangod/Aql/AqlItemMatrix.h
@@ -95,6 +95,7 @@ class AqlItemMatrix {
   size_t numberOfBlocks() const noexcept;
 
   std::pair<SharedAqlItemBlockPtr, size_t> getBlock(size_t index) const noexcept;
+  std::pair<AqlItemBlock const*, size_t> getBlockRef(size_t index) const noexcept;
 
   bool stoppedOnShadowRow() const noexcept;