@Dronplane 2021-03-12
Yes, these changes just removes SharedBlockPtr copying/creating overhead (mostly reference counting). This places were dark red in performance profiler :(.